O dia todo tentando colocar uma maldita mensagem de confirmacao

+++ src/main/java/br/uem/gestaoresiduos/services/LaboratorioService.java
@@ -30,7 +30,7 @@ public class LaboratorioService {
 		return laboratorioRepository.findByBloco(bloco);
 	}
 	
-	public Laboratorio findLaboratorioById(Long id) {
+	public Laboratorio findLaboratorioById(int id) {
 		return laboratorioRepository.findOne(id);
 		
 	}
